Reorganized the API endpoints into separate modular components using express.Router. will make it much easier to read and implement new endpoints
